Saving, Editing, and Deleting Recipes is Now Possible
In this commit, recipes can now be saved, loaded, and deleted. Swipe to delete and an edit button allow for recipe deletion. Selecting the edit button on a recipe allows for it to be edited. You can selected multiple ingredients for a recipe and those selected ingredients will get added to an ingredients array in the IngredientSelectionViewModel, but this array doesn't do anything yet. Conversely, deselecting an ingredient will remove it from that array. However, selected ingredients will not be checked off in the IngredientSelectionViewController yet, I have not yet implemented this functionality.