Adding Ingredients Updates AddEditRecipeViewController's TableView
I've been struggling with a bug wherein the tableview in the AddEditRecipeViewController does not update its cell that displays the recipe's ingredients after they've been selected. Before, in order for this View Controller to display the selected ingredients, you had to select the ingredients, save the recipe, and then reopen it. Now the ingredients are displayed after they're selected.