Implemented Search, Data Generation, and Realtime UI Updates
In this commit, search can filter by all Recipe data except for Ingredients, as I still have to figure out how to implement that.

The recipe categories and ingredients that come stock with the app now get generated only once, which prevents the user from generating duplicate categories and ingredients every time they run the app. I did this with UserDefaults.

Finally, this commit implements changes to the AddEditRecipeViewController that allows the selected ingredients cell to update immediately after ingredients are selected from the IngredientSelectionViewController. Before, the AddEditRecipeViewController had to be closed and reopened before the list of selected ingredients would update within the table view.