RecipeDetailsViewController and Bug Fixes
This commit features a RecipeDetailsViewController that displays details about the recipe that corresponds its Details button is tapped in the HomeViewController. It also resolves a bug wherein text would be deleted from the textfield of an AddEditRecipeTableViewCell when the cell was dequeued. This bug persists with the picker and ingredients list in the AddEditRecipeViewController, this will be resolved in a future commit.